GUI: Belli aralıklarla zamanlanmış zincir oluşturmada çıkan hata handle edildi

+++ DiskBackup.Business/Abstract/IBackupService.cs
@@ -33,7 +33,7 @@ namespace DiskBackup.Business.Abstract
 
         //INC ve DIFF TaskInfo'dan strObjeyi alarak çoklu seçim yapabilir
         [OperationContract]
-        byte CreateIncDiffBackup(TaskInfo taskInfo);
+        int CreateIncDiffBackup(TaskInfo taskInfo);
         [OperationContract]
         bool CreateFullBackup(TaskInfo taskInfo); //Bu daha hazır değil
